requirements_v2.txt
# RAEC v2 Requirements

# Discord Bot Framework
discord.py>=2.3.0

# Google Gemini AI (New Package)
google-genai>=0.1.0

# Environment Variables
python-dotenv>=1.0.0

# Database (Built-in with Python)
# sqlite3 - No installation needed
